Add option for immediate expansion
Currently, fluidbox waits for the high-res image to load before expanding.
If you're linking to very large images, this leads to a laggy UI and may
confuse the user into thinking that the link is broken. This adds an option
$.fluidbox({immedate:true}) which will trigger the lightbox load immediately
on click and begin loading the high-res image in the background. When the
high-res has loaded, the low-res is replaced. This commit also adds a utility
class 'fluidbox-loading' and an empty 'fluidbox-status' div on the ghost object
to allow the user to style it with their own UI cues.