Updated to version v1.4.3
Enabled the option to trigger immediate opening of Fluidbox. This has
to be turned on manually, by setting `immediateOpen` to  `true` during
initialization.

Added five new custom events that are triggered.

Rearranged some code with regards to opening and closing of Fluidbox
instances for better readability, fixed indentation inconsistencies in
original JS file.

Added an animated CSS loader to Fluidbox instances, exceptionally
important for delayed opening setups (default setup) as users are often
unaware if Fluidbox has failed, or simply waiting for target image to
be preloaded. The appearance of the loader is purely controlled by a
CSS pseudo-element, activated upon the addition of `.fluidbox-loading`
class.

The stylesheet of Fluidbox has been ported over to LESS, in addition to
SCSS.