Pre-v1.4.4 updates
Removed .less file, because SCSS itself is sufficient and it is tiring
to maintain two pre-processed stylesheets (which defeats the purpose at
the first place).

Bug fix for Safari, which is extremely stubborn with transformed
elements (with background images) that are overlaid with absolutely
positioned child elements (be it pseudo-element or otherwise). Might
consider rolling back to a simpler solution once Safari stops acting up.

Also updated demo to demonstrate the rendered markup after Fluidbox
initialisation.