Enabled URL checking and cleaning
Fluidbox will throw a warning and refuse to proceed if whitespace
characters are present in the URL. As HTML collapses multiple
whitespaces into one, it is impossible to determine whether the
reference file contains collapsed or uncollapsed multiple whitespaces.
Whitespaces must be escaped manually.

On the other hand, Fludibox throws a warning but proceeds with expected
behaviour when encountering the following character set (expressed in
regex): ["'\(\)]. The reason being that these characters directly
interfere with jQuery and therefore need to be replaced anyway.